added SSHADER_S header file
added new way of adding shader files to OpenGL context.
in main, we create object of Shader class found in the header file, the constructor in header file reads the file and stores it in a string, then the vertex and fragment objects are created in the constructor. compiled and then a shader program combines them, compiles & link and stores it in a  ID